requirements.txt
beautifulsoup4==4.13.4
Requests==2.32.4
yt_dlp==2025.6.25
python-dotenv==1.1.1
